Ensure referrer is sent client side on opera mini
